% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_orderings.R
\name{adjacency_to_paths}
\alias{adjacency_to_paths}
\title{Given adjacency matrix input,
returns a list of simple paths that are not a strict subset of another simple path.}
\usage{
adjacency_to_paths(adj_mat)
}
\arguments{
\item{adj_mat}{Adjacency matrix output from \code{column_ordering_matrix()}.}
}
\description{
Given adjacency matrix input,
returns a list of simple paths that are not a strict subset of another simple path.
}
\keyword{internal}
