% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_identical}
\alias{all_identical}
\title{Test several objects are all identical}
\usage{
all_identical(...)
}
\arguments{
\item{...}{Objects to compare.}
}
\value{
Logical. \code{TRUE} if and only if all objects are \code{identical()}.
}
\description{
Test several objects are all identical
}
