% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_by_index.R
\name{case_by_index}
\alias{case_by_index}
\title{Choose values row-wise from a set of columns according to an index column}
\usage{
case_by_index(data, index_column, ..., into = "output")
}
\arguments{
\item{data}{A data frame.}

\item{index_column}{A column of integers indicating which column to select from for each row.}

\item{...}{Columns to select values from. Passed to \code{dplyr::select} so can use tidy-select.}

\item{into}{String giving the name of the new column.}
}
\value{
A data frame which matches the input \code{data} with a new column.
}
\description{
Adds a new column to the input \code{data}. The new column takes values row-wise from a supplied
subset of columns in \code{data}. Which column's value is used for each row is determined by the
corresponding integer in the column in the data given by \code{index_column}.
}
\examples{
x <- tibble::tibble(a = 11:13, b = 21:23, index = c(1, 2, 1))
case_by_index(x, index, a, b, into = "new")
}
