% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowrun_dice.R
\name{colorize_d6}
\alias{colorize_d6}
\title{Colorize a diceroll}
\usage{
colorize_d6(throw, collapse = TRUE)
}
\arguments{
\item{throw}{Integer vector, like \code{sr_d6(5)$throw}}

\item{collapse}{Logical. Whether to concatenate the integers.}
}
\value{
Integer vector or \code{character(1)} depending on \code{collapse}.
}
\description{
Colorize a diceroll
}
\keyword{internal}
