% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_orderings.R
\name{column_orderings}
\alias{column_orderings}
\title{Get sets of column orderings}
\usage{
column_orderings(df)
}
\arguments{
\item{df}{A data frame with only atomic columns (no list or data frame columns).}
}
\value{
A list of vectors of column names. Each name vector gives an ordering of columns
largest to smallest.
}
\description{
Column A is ordered ahead of column B if \code{A[i]>=B[i]} for all rows i in which neither A or B has
missing values. A column ordering is a sequence of column names which are in order e.g. A>=B>=C.
This function returns a list of all \emph{possible} orderings which are not a subset of another
ordering. The orderings are possible not definite where there are \code{NA}s (treated as missing values).
}
