% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_elements.R
\name{equal_elements}
\alias{equal_elements}
\title{Get sets of equal elements}
\usage{
equal_elements(x, keep_names = TRUE)
}
\arguments{
\item{x}{A list or atomic vector.}

\item{keep_names}{Logical. If \code{TRUE} and \code{x} has names then returned sets will be names from \code{x},
otherwise indices.}
}
\value{
A list of either names or indices of \code{x} (see \code{keep_names}).
}
\description{
Divides \code{x} into disjoint subsets based on equality between elements. Equality between
pairs of elements is determined by \code{all.equal()}.
}
