% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_version.R
\name{get_rs_version}
\alias{get_rs_version}
\title{Get Current RStudio Version}
\usage{
get_rs_version()
}
\value{
A \code{character} of length 1
}
\description{
Get Current RStudio Version
}
\examples{
\dontrun{
get_rs_version()
}
}
