% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-helpers.R
\name{mattex}
\alias{mattex}
\title{Print a matrix for LaTeX.}
\usage{
mattex(mat, type = "pmatrix", round = 3)
}
\arguments{
\item{mat}{A \link{matrix}.}

\item{type}{\verb{[character(1): "pmatrix"]}: Matrix environment to use, e.g. \code{bmatrix} for \verb{[]}-enclosed
(brackets) or \code{pmatrix} for \verb{()} (parentheses).}

\item{round}{\verb{[integer(1): 3]}}
}
\value{
Nothing, output is \code{cat}ed.
}
\description{
Print a matrix for LaTeX.
}
\examples{
mat <- matrix(1:9, ncol = 3)
mattex(mat)
}
