% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blogdown-new-post.R
\name{new_post}
\alias{new_post}
\title{Create a new blogdown post (with tweaks)}
\usage{
new_post(
  title,
  kind = "",
  open = interactive(),
  author = getOption("blogdown.author"),
  categories = NULL,
  tags = NULL,
  packages = NULL,
  shows = NULL,
  date = Sys.Date(),
  file = NULL,
  slug = NULL,
  title_case = getOption("blogdown.title_case"),
  subdir = getOption("blogdown.subdir", "post"),
  ext = getOption("blogdown.ext", ".md")
)
}
\arguments{
\item{title}{The title of the post.}

\item{kind}{The archetype to use.}

\item{open}{Whether to open the post in the editor, defaults to \code{TRUE} if in
an interactive setting.}

\item{author}{The author of the post.}

\item{categories}{A character vector of category names.}

\item{tags}{A character vector of tag names.}

\item{packages, shows}{A character vector of package names or show names.}

\item{date}{The date of the post.}

\item{file}{The filename of the post. By default, the filename will be
automatically generated from the title by replacing non-alphanumeric
characters with dashes, e.g. \code{title = 'Hello World'} may create a file
\file{content/post/2016-12-28-hello-world.md}. The date of the form
\code{YYYY-mm-dd} will be prepended if the filename does not start with a
date.}

\item{slug}{The slug of the post. By default (\code{NULL}), the slug is
generated from the filename by removing the date and filename extension,
e.g., if \code{file = 'post/2015-07-23-hi-there.md'}, \code{slug} will be
\code{hi-there}. Set \code{slug = ''} if you do not want it.}

\item{title_case}{A function to convert the title to title case. If
\code{TRUE}, the function is \code{tools::\link[tools]{toTitleCase}()}).
This argument is not limited to title case conversion. You can provide an
arbitrary R function to convert the title.}

\item{subdir}{If specified (not \code{NULL}), the post will be generated
under a subdirectory under \file{content/}. It can be a nested subdirectory
like \file{post/joe/}.}

\item{ext}{The filename extension (e.g., \file{.md}, \file{.Rmd}, or
\file{.Rmarkdown}). Ignored if \code{file} has been specified.}
}
\description{
This is a modified version of \code{blogdown::new_post} for extra taxonomies.
}
