% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-reimport.R
\name{pkg_export}
\alias{pkg_export}
\title{Export installed packages for somehwat functional reinstallation}
\usage{
pkg_export(old_lib = "~/Library/R/shared_bak/")
}
\arguments{
\item{old_lib}{\code{character}: The location of your previous package library.}
}
\description{
Export installed packages for somehwat functional reinstallation
}
\examples{
\dontrun{
pkg_export()
}
}
