% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowrun_dice.R
\name{print.sr_d6}
\alias{print.sr_d6}
\title{Print diceroll}
\usage{
\method{print}{sr_d6}(x, ...)
}
\arguments{
\item{x}{Object of class \code{sr_d6}, like from \link{sr_d6}.}

\item{...}{Ignored}
}
\value{
Nothing
}
\description{
Print diceroll
}
\examples{
sr_d6(5)
}
