% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-helpers.R
\name{render_tex_files}
\alias{render_tex_files}
\title{Render a bunch of small .tex files}
\usage{
render_tex_files(
  path,
  cleanup = TRUE,
  tex_fun = tinytex::xelatex,
  to_png = TRUE
)
}
\arguments{
\item{path}{Folder where \code{.tex} files are.}

\item{cleanup}{\code{TRUE}: Whether to cleanup auxilliary files (e.g. \code{.aux}, \code{.log}, ...)}

\item{tex_fun}{Which function to use, default ist \link[tinytex:xelatex]{tinytex::xelatex}.}

\item{to_png}{\code{TRUE}: Whether to use \code{convert} to convert PDF to png.}
}
\value{
Nothing
}
\description{
Use case: A bunch of stuff done in LaTeXiT, resulting in some helper calculations
or derivations contained in tex files. These are then rendered to PDF using \code{tex_fun}, per
default \link[tinytex:xelatex]{tinytex::xelatex}, and then optionally converted to \code{png} using the system \code{convert}
tool via imagemagick.
}
\examples{
\dontrun{
render_tex_files("formulas")
}
}
