% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_version.R
\name{rs_daily}
\alias{rs_daily}
\title{Get Current RStudio Daily Version}
\usage{
rs_daily()
}
\value{
A \code{character} of length 1
}
\description{
Get Current RStudio Daily Version
}
\examples{
rs_daily()
}
