% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_version.R
\name{rs_preview}
\alias{rs_preview}
\title{Get Current RStudio Preview Version}
\usage{
rs_preview()
}
\value{
A \code{character} of length 1
}
\description{
Get Current RStudio Preview Version
}
\examples{
rs_preview()
}
