% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_version.R
\name{rs_release}
\alias{rs_release}
\title{Get Current RStudio Version}
\usage{
rs_release()
}
\value{
A \code{character} of length 1
}
\description{
Get Current RStudio Version
}
\examples{
rs_release()
}
