% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowrun_dice.R
\name{sr_d6}
\alias{sr_d6}
\title{Roll some d6 for Shadowrun}
\usage{
sr_d6(n)
}
\arguments{
\item{n}{Number of d6 to roll}
}
\value{
A \code{list} of additional class \code{"sr_d6"}.
}
\description{
Roll some d6 for Shadowrun
}
\examples{
sr_d6(10)
}
