% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_xtab.R
\name{tidyxtab}
\alias{tidyxtab}
\title{Tabularize tidily}
\usage{
tidyxtab(data, x, y, row_name = deparse(x), col_name = deparse(y))
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{x}{The row-variable}

\item{y}{The column-variable}

\item{row_name, col_name}{Passed to \link[janitor:adorn_title]{janitor::adorn_title}.}
}
\value{
A \link[janitor:tabyl]{janitor::tabyl}
}
\description{
Use \link[janitor:tabyl]{janitor::tabyl} and various \code{adorn_}-functions to create a tidy crosstab.
}
\examples{
tidyxtab(mtcars, carb, gear)
}
