% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_spec.R
\name{col_spec}
\alias{col_spec}
\title{Get column types for use as \code{col_types} in \code{readr::read_csv()}}
\usage{
col_spec(data)
}
\arguments{
\item{data}{A data frame.}
}
\description{
Returns the compact string representation for the \code{col_types} argument in
\code{readr::read_csv()} and related functions of the supplied \code{data}. The intended
use is to record the types after an initial load (with any further processing), so it
can be used for future reads of the same data.
}
\details{
The column types will be one of:
\itemize{
\item c = character
\item i = integer
\item d = double
\item l = logical
\item f = factor
\item D = date
\item T = date time
\item t = time
}

These types are read directly from \code{data} - there is no guessing or coercion.
The first four are identified using \code{typeof()}. The remainder overwrite the type if
one of the following is true for the column: \code{is.factor()}, \code{lubridate::is.Date()},
\code{lubridate::is.POSIXt()}, or \code{"hms" \%in\% class(.)}.
}
