% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolors.R
\name{coolors_pal}
\alias{coolors_pal}
\title{Get a color palette from a coolors.co URL}
\usage{
coolors_pal(url)
}
\arguments{
\item{url}{The link to a coolors.co palette, e.g.
\code{"https://coolors.co/29adb2-182326-2a3a3f-69818c-8ea4b2"}}
}
\value{
\code{character(5)} with RGB color codes, e.g.
\code{c("#29adb2", "#182326", "#2a3a3f", "#69818c", "#8ea4b2")} for the example \code{url}.
}
\description{
Get a color palette from a coolors.co URL
}
\note{
It's only the URL that's parsed, there's no need for an internet connection.
The \code{url} does not need to be valid, since only hex strings of length 6 (i.e.
color codes like \code{F75AB1}) are looked for, and a \verb{#} is prepended.
}
\examples{
coolors_pal("https://coolors.co/29adb2-182326-2a3a3f-69818c-8ea4b2")
}
