% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jemisc-package.R
\docType{package}
\name{jemisc-package}
\alias{jemisc}
\alias{jemisc-package}
\title{jemisc: Some miscellaneous functions for personal use}
\description{
A package for functions that are speculative or only used infrequently.
}
\author{
\strong{Maintainer}: James Edwards \email{jamese1326@gmail.com}

}
\keyword{internal}
