% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jemisc-package.R
\docType{package}
\name{jemisc-package}
\alias{jemisc}
\alias{jemisc-package}
\title{jemisc: Personal Collection of Stuff and Things}
\description{
Just another personal R package.
}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@quantenbrot.de}

}
\keyword{internal}
