% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jemisc-package.R
\docType{package}
\name{jemisc-package}
\alias{jemisc}
\alias{jemisc-package}
\title{jemisc: Personal Collection of Stuff and Things}
\description{
Just another personal R package. This package
    provides no greater value to the R community, it is merely a place for
    me to dump stuff that I ended up writing and found mildly useful.
}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@quantenbrot.de}

}
\keyword{internal}
